created WorldMapPermissionHelper.java to move checks out of the DatasetPage. #2290
